+++ stage_3_sources/stdio/ostream.h
@@ -13,7 +13,7 @@ namespace std
 		ostream (std::streambuf<char> &buff) : buff(buff) {}
 
 		void putString (std::string str) {
-			for (int i = 0; i < str.size(); i++)
+			for (size_t i = 0; i < str.size(); i++)
 				buff.put(str[i]);
 		}
